requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
datasets
semantic-router
python-dotenv
yandex-cloud-ml-sdk
lxml
langchain_community
nltk
bs4
requests